Recover ~28 KB of boot-resident internal DRAM for the portal serving budget
The web portal exhausted internal heap serving a WiFi client (ISS-192):
AP joinable, DHCP flaky, HTTP dead. Move occasional-feature buffers out
of always-resident DRAM and bound the portal window's runtime costs:

- Sync payload staging (8.2 KB) and screencap base64 (1.4 KB) allocate
  on demand (PSRAM-first) and free when idle
- Live-caption session buffers (7.7 KB) live only while a live session
  runs; allocation failure refuses the session gracefully
- Voice recorder list/drain buffers (7.7 KB) are app-scoped
- CRC32 table becomes constexpr in flash; appDefs becomes const
- Serial RX ring 16 KB -> 8 KB (sync is stop-and-wait, 4 KB max burst)
- AsyncTCP task stack 16 KB -> 8 KB, WS queue/client caps bounded
- mDNS deferred until STA connects (unused in AP-only onboarding)
- I2S TX released during the portal session (DMA-capable heap back)
- Zero-copy portal page send (also fixes silent empty-200 on OOM)
- MatrixScreensaver/Starburst structs shrunk to bounded int types

Static DRAM: 115,884 -> 87,820 bytes (linker map, local env).
Native suites 124/124; wasm emulator + device-module contract + six
core device builds green.

+++ lib/MatrixScreensaver/MatrixScreensaver.h
@@ -59,8 +59,8 @@ private:
     // Info about each row
     struct RowInfo {
         char ch;
-        int litPixels;          // 0..FONT_HEIGHT
-        unsigned long nextChange; // next time to flicker
+        uint8_t litPixels;      // 0..FONT_HEIGHT (8)
+        uint32_t nextChange;    // next time to flicker
     };
 
     struct Column {
